requirements.txt
fastapi
uvicorn
python-dotenv
bcrypt
pandas
pydantic[email]
python-multipart 
pandantic
fpdf2 
coverage
httpx